psc: fix dependency argument on init

+++ nx/include/switch/services/psc.h
@@ -94,7 +94,7 @@ void pscmExit(void);
 /// Gets the Service object for the actual psc:m service session.
 Service* pscmGetServiceSession(void);
 
-Result pscmGetPmModule(PscPmModule *out, PscPmModuleId module_id, const u16 *dependencies, size_t dependency_count, bool autoclear);
+Result pscmGetPmModule(PscPmModule *out, PscPmModuleId module_id, const u32 *dependencies, size_t dependency_count, bool autoclear);
 
 Result pscPmModuleGetRequest(PscPmModule *module, PscPmState *out_state, u32 *out_flags);
 Result pscPmModuleAcknowledge(PscPmModule *module, PscPmState state);
